Refactoring of TraceReader to use hooks
Events comparison with EventQuery inside switch case are cumbersome and offer
poor performance. We replaced them by hooks mechanism about the same way as
LTTV. Each ITraceEventHandler must create and add hooks it needs. A function
whose name is the concatenation of "handle_" + traceFileName + "_" eventName
that accept TraceReader and JniEvent as parameters must be defined. If a
TraceHook is created without argument, the function "handle_all_event" will be
called instead for each events in the trace.