Add ant build system and required dependencies
As of now, a modified version of swtchart is required to run flightbox. We
added the source here because upstream didn't merged the required patch.

The jar-all target creates an archive with all dependencies resolved into one
convenient package. The jni trace read library is still needed.

The junit target is broken at this moment.