Recover sockets and related addresses from tasks
Recovery of sockets is done according to syscalls. The source and destination
addresses are recovered for client sockets inside the connect system call. For
server, the corresponding addresses of the socket is obtained from the softirq
that triggers the wakeup.

The algorithm support only simple accept and make two assumptions:
  * connect will always have their corresponding dev_xmit inside the system call
  * accept is always triggered by wakeup in the receiving softirq, event if the
    task is already running