Handle subtask blocking resolution somewhere else
It's not possible to reliably add all subblocking items to a task in the case
of sockets. The relation between the socket fd and the addresses are known only
near the exit of sys_accept, after the wait is poped from the state stack.
Since all blockings are known at the report generation time, they can be
computed at the end of the system call, or even at the end of the trace
processing.